fix: Generate report XLSX.
